Liking posts and comments

+++ src/main/scala/na/przypale/fitter/repositories/CommentsCountersRepository.scala
@@ -3,7 +3,7 @@ package na.przypale.fitter.repositories
 import na.przypale.fitter.entities.Comment
 
 trait CommentsCountersRepository {
-  def getLikesAndAnswers(comment: Comment):  Map[String, Int]
+  def getLikesAndAnswers(comment: Comment):  Map[String, Long]
   def likeComment(comment: Comment)
   def increaseAnswersCounter(comment: Comment)
   def initiateCounters(comment: Comment)
