+++ src/main/scala/na/przypale/fitter/repositories/PostsCountersRepository.scala
@@ -3,7 +3,7 @@ package na.przypale.fitter.repositories
 import na.przypale.fitter.entities.Post
 
 trait PostsCountersRepository {
-  def getLikesAndComments(post: Post):  Map[String, Int]
+  def getLikesAndComments(post: Post):  Map[String, Long]
   def likePost(post: Post)
   def increaseCommentsCounter(post: Post)
   def initiateCounters(post: Post)
